Added functionalities in video controller

+++ src/controllers/video.controller.js
@@ -4,6 +4,9 @@ import { ApiResponse } from "../utils/ApiResponse.js";
 import { uploadOnCloudinary } from "../utils/cloudinary.js";
 import { Video } from "../models/video.models.js";
 import mongoose from "mongoose";
+import { User } from "../models/user.models.js";
+import {Comment} from "../models/comment.model.js"
+import {Like} from "../models/like.model.js"
 
 const getAllVideos = asyncHandler(async (req, res) => {
     const {page=1, limit=10, query, sortBy, sortType, userId}= req.query
@@ -111,16 +114,44 @@ const getAllVideos = asyncHandler(async (req, res) => {
                     }
                 ]
             }
+        }, 
+        {
+            $sort: {
+                [sortBy]: sortType === "desc" ? -1 : 1,
+                createdAt: -1
+                //Sort by createdAt in descending order as an option newest first
+            } //Sort by ascending order (1) or descending order (-1)
+        }, 
+        {
+            $skip: skip
+            //Skip documents for pagination
+        },
+        {
+            $limit: pageLimit
         }
       ]
 
+      console.log(pipeline, 'Pipeline of videos');
+      if(!pipeline || pipeline.length === null){
+        throw new ApiError(500, "Loading Failed, Please try again later")    
+      }
 
+      const video = await Video.aggregate(pipeline)
+      const videoPaginate = await Video.aggregatePaginate(pipeline)
 
+      if(!video || video.length==(0 || null)){
+        throw new ApiError(500, "Failed to get all videos, please try again later")
+      }
+
+      return res
+      .status(200)
+      .json( new ApiResponse(200, {video, videoPaginate}, "Video Pagination and Aggregation Retrived Successfully"))
 })
 
 const publishVideo =  asyncHandler(async(req, res)=>{
     const {title, description} = req.body
-    if(!title && !description){
+    if(!title || !description){
+        //If any one of them is absent, throw the error
         throw new ApiError(404, "Title and description is required")
     }
     let videoUploadLocal = req.files?.videoUpload?.[0]?.path
@@ -151,7 +182,8 @@ const publishVideo =  asyncHandler(async(req, res)=>{
 })
 
 const updateVideoDetails =  asyncHandler(async(req, res)=>{
-    const {videoId} = req.params
+    const {videoId} = req.params 
+    //Destructuring should be same as in defined route name
     const {title, description} =  req.body
     const existedTitle= await Video.findOne(
         {title: title}
@@ -187,10 +219,30 @@ const getVideosById = asyncHandler(async(req, res)=>{
     if(!videoId){
         throw new ApiError(500, 'Id not found')
     }
+
+    const user = await User.findById(req.user._id)
     const video = await Video.findById(videoId)
     if(!video){
         throw new ApiError(404, 'Video not found')
     }
+
+    if(!video.isPublished){
+        throw new ApiError(403, "Video is not published yet")
+    }
+
+    const updateVideoInfo = await Video.updateOne(
+        {_id: videoId},
+        {
+            $inc: {
+                views: 1
+            }
+        },
+        { new: true, validateBeforeSave: false}      
+    )
+    if(updateVideoInfo.nModified===0){
+        throw new ApiError(404, "Video not Found, no updates happened")
+    }
+
     return res.status(200)
     .json(
         new ApiResponse(200, video, 'Video and its details found successfully')
@@ -207,15 +259,67 @@ const deleteVideo = asyncHandler(async(req, res)=>{
         throw new ApiError(404, 'Video not found')
     }
 
-    const videoDelete =  await Video.findByIdAndDelete(videoId)
-    //Returns the deleted document
+    const thumbnailURL = video.thumbnail
+    //Extract video URL from video document
+    const urlArrayofThumbnail = thumbnailURL.split("/")
+    const thumbnailfromURL = urlArrayofThumbnail[urlArrayofThumbnail.length-1]
+    const thumbnailName = thumbnailfromURL.split(".")[0]
+
+    //Deleting video from database, then from cloudinary
+    if(video.owner.toString()=== req.user._id.toString){
+        const videoDeleteFromDatabase =  await Video.findByIdAndDelete(videoId)
+        if(!videoDeleteFromDatabase){
+            throw new ApiError(404, "Video Not Found, might be already deleted")
+        }
+
+        //Delete Video File from Cloudinary
+        await deleteOnCloudinaryVideo(video.videoFile)
+
+        //Deleting thumbnail from Cloudinary
+        await cloudinary.uploader.destroy(thumbnailName,
+            {
+                invalidate: true
+            },
+            
+                (error, result)=>{
+                    console.log("Result:", result, "Error:", error, "Result or Error after deleting thumbnail from cloudinary");
+                }
+            )
+        const comments = await Comment.find({video: videoDeleteFromDatabase._id})
+        //Taking out the comment IDs
+        const commentIDs = comments.map((comment)=>{comment._id})
+        //deleteMany()-->Removes all the documents that match the filter from a collection
+        await Like.deleteMany({video: videoDeleteFromDatabase._id})
+        await Like.deleteMany({comment: { $in: commentIDs}})
+        await Comment.deleteMany({video: videoDeleteFromDatabase._id})
+    } else {
+        throw new ApiError(403, "You are not authorized to delete this video")
+    }
+
+    
+    //Return the deleted document
     return res.status(200)
     .json( new ApiResponse(200, videoDeleted, "Deleted Successfully"))
 })
 
-const togglePublishStatus = asyncHandler(async(req,re)=>{
+const togglePublishStatus = asyncHandler(async(req,res)=>{
+    //To toggle between published status==> On or Off
     const {videoId} =  req.params
+    if(!videoId){
+        throw new ApiError(404, "Invalid Video ID to know publish status")
+    }
+    const video =  await Video.findById(videoId)
+    if(!video){
+        throw new ApiError(400, "Cannot toggle publish status, either video does not exist or deleted")
+    }
+
+    video.isPublished = !video.isPublished
+    await video.save({validateBeforeSave: false})
+
+    return res.status(200)
+    .json(new ApiResponse(200, videoId, "Video Status is toggled successfully"))
 })
+//If isPublished is true video will be shown, otherwise not
 
 
 export {
